Clean-up forward declarations and #include around sycl::vec<>
